# [do not merge] testing rotary embedding + torch.compile