@cpuhrsch 2024-08-31
I'd call `pipe` a few times (say 5) within `run_inference` to make sure warmup and initial setup cost isn't being captured.